+++ src/tokens/components/error-page.ts
@@ -1,0 +1,4 @@
+import errorPage from '../../../tokens/component/error-page.json';
+import { flattenTokens } from '../utils';
+
+export const errorPageTokens = flattenTokens(errorPage as never);